+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- [TD-5495] Foreing keys columns should match original ID columns in all tables
+
 ## [6.3.0] 2024-03-18
 
 ### Added